+++ tests/trainer/optimization/test_multiple_optimizers.py
@@ -45,7 +45,7 @@ def test_unbalanced_logging_with_multiple_optimizers(tmpdir):
 
     # Initialize a trainer
     trainer = pl.Trainer(
-        default_root_dir=tmpdir, max_epochs=1, limit_train_batches=5, limit_val_batches=5, weights_summary=None
+        default_root_dir=tmpdir, max_epochs=1, limit_train_batches=5, limit_val_batches=5, enable_model_summary=False
     )
     trainer.fit(model)
 
@@ -77,7 +77,7 @@ def test_multiple_optimizers(tmpdir):
         limit_val_batches=2,
         max_epochs=1,
         log_every_n_steps=1,
-        weights_summary=None,
+        enable_model_summary=False,
     )
     trainer.fit(model)
 
@@ -117,7 +117,7 @@ def test_multiple_optimizers_manual(tmpdir):
     model.val_dataloader = None
 
     trainer = pl.Trainer(
-        default_root_dir=tmpdir, limit_train_batches=2, max_epochs=1, log_every_n_steps=1, weights_summary=None
+        default_root_dir=tmpdir, limit_train_batches=2, max_epochs=1, log_every_n_steps=1, enable_model_summary=False
     )
     trainer.fit(model)
 
@@ -189,7 +189,7 @@ def test_custom_optimizer_step_with_multiple_optimizers(tmpdir):
         limit_train_batches=limit_train_batches,
         max_epochs=1,
         log_every_n_steps=1,
-        weights_summary=None,
+        enable_model_summary=False,
     )
     trainer.fit(model)
     assert len(model.training_step_called) == len(model.optimizer_step_called) == len(model.optimizers())